requirements.txt
Flask==2.3.2
supabase==0.3.0
python-dotenv==1.0.0
Flask-Cors==3.0.10
